requirements.txt
python-dotenv
paho.mqtt